move check for current group from home page to login page
