+++ android/app/src/main/AndroidManifest.xml
@@ -12,8 +12,7 @@
 
         <activity
             android:name=".userRelated.FriendList"
-            android:label="@string/friendList"
-            android:parentActivityName=".ui.firstLaunch.Home_page"></activity>
+            android:label="@string/friendList"></activity>
         <activity android:name=".userRelated.SettingsPage" />
         <activity
             android:name=".ui.firstLaunch.ResetPassword"
@@ -21,12 +20,10 @@
             android:parentActivityName=".ui.firstLaunch.LoginActivity" />
         <activity
             android:name=".everythingWithGroups.CreateGroupPage"
-            android:label="@string/title_create_group"
-            android:parentActivityName=".ui.firstLaunch.Home_page" />
+            android:label="@string/title_create_group" />
         <activity
             android:name=".everythingWithGroups.InsideGroupActivity"
             android:label="@string/title_activity_inside_group"
-            android:parentActivityName=".ui.firstLaunch.LoginActivity"
             android:theme="@style/AppTheme.NoActionBar" />
         <activity android:name=".everythingWithGroups.JoinAGroupById"/>
         <activity android:name=".ui.firstLaunch.Home_page" />
